uint32 -> uint64

+++ src/random.h
@@ -408,7 +408,7 @@ INLINE static void random_direction_in_cone(const int64_t id_bh,
 INLINE static double random_gaussian(const double mean, const double std) {
 
   /* Generate two uniform random numbers for sampling the gaussian. */
-  static uint32_t seed = 42;
+  static uint64_t seed = 42;
   uint16_t xsubi[3] = {seed & 0xFFFF, (seed >> 16) & 0xFFFF,
                        (seed >> 32) & 0xFFFF};
   double u1 = inl_erand48(xsubi);
